+++ tests/Browser/Nova/PersonbydomainsTable/Policies/Lookup_roles/Attach/AttachAnyWhenARoleExistsTest.php
@@ -78,7 +78,7 @@ class AttachAnyWhenARoleExistsTest extends PersonbydomainsTableBaseDuskTest
                 ->pause($pause['long'])
                 ->assertVisible('@2-view-button')
                 ->click('@2-view-button')
-                ->pause($pause['short'])
+                ->pause($pause['long'])
                 ->assertSee('Personbydomain Details')
                 ->assertSee('Lookup User Role')
                 ->assertVisible('@2-row')
